First commit of OpenGL::Modern
This is a rough start for the POGL2 development.
The module is named OpenGL::Modern and while it should
eventually support legacy OpenGL 1.x, 2.x, and 3.0 APIs,
the focus is support for the new shader function pipeline.

Thanks to Max Maischein for his work implementing app-shadertoy
which showed us what modern OpenGL can do from perl and whose
OpenGL::Glew autogenerated bindings form the start for this
code base.