Remove app-shadertoy utils from OpenGL::Modern distribution
Per discussion on the pogl-general mailing list
and as noted in the OpenGL-Modern wiki at
https://github.com/devel-chm/opengl-modern/wiki
the OpenGL::Modern::OpenGL4, OpenGL::Modern::Helpers,
and OpenGL::Texture are all modules that will need to
be coordinated with the existing OpenGL::Shader and
OpenGL::Image modules for user transition.  The
Helper.pm is not needed since the API is now being
implemented directly in OpenGL::Modern.  Experiments
can be done via branching.