Fix glewCreateContext and glewInit return values
Added a return value for glewCreateContext() and note
that GLEW_OK is 0 (the same as GL_FALSE) which may be
why glewCreateContext() returns 0 on apparent success.
It looks like we should be able to use these two
routines, check their error status, and determine
whether or not to skip tests for OpenGL::Modern.
With luck, this will fix some of the FAILs from CPAN
Testers since they appear to be from systems without
OpenGL Context support.