Add implementation TODO and NOTE for Helpers.pm
For consistency, I propose to rename the _p routines in OpenGL
involving OpenGL::Array objects as _o for _oga.  This is a change
from OpenGL but it is clear about the usage and better fits a
general naming scheme for OpenGL::Modern bindings that are not
restricted to only one preferred implementation or data type.