MacOSX: allow build with homebrew and its emulated MesaGL
This change comes from Prima::OpenGL that is build with Prima that uses lots of
homebrew libraries, and is actually advised to be build on homebrew. The
problem comes with the libraries that are implicitly dependant on the native
/opt/X11 libraries, and any program that is compiled with both set of Xlibs
(and by extention, GL libs), dumps core.

A possibile solution to this is to let OpenGL::Modern to use homebrew libraries,
but that comes with a price, native mac gl will not be used, but the slow
emulated MesaGL will be used instead.